+++ src/main/java/csc1035/project2/MainIO.java
@@ -185,12 +185,50 @@ public class MainIO {
                     "5: Return to main menu.\n");
             int choice = sc.nextInt();
             switch(choice){
+                case 1 -> studentAddModule();
                 case 5 -> {
                     System.out.println("Returning to main menu.");
                     quit = true;
                 }
             }
             default -> System.out.println("Not a valid option.");
+        }
+    }
+
+    // TODO: Confirm if we want an updating Room/Module Handler or create a new one each option pick.
+    private void studentAddModule() {
+        Scanner sc = new Scanner(System.in);
+        IController ic = new Controller<>();
+        ModuleHandler handler = new ModuleHandler();
+
+        System.out.print("Please enter Student ID: ");
+        String sid = sc.nextLine();
+        try {
+            Students student =  (Students) ic.getById( Students.class, sid );
+
+            System.out.print("Which module would you like the student to take?\n" + "Please enter Module ID");
+            String mid = sc.nextLine();
+            try {
+                Modules modules = (Modules) ic.getById( Modules.class, mid );
+                boolean added = handler.addStudentToModule( student, modules );
+                if ( added ) {
+                    System.out.println("Addition successful.\n" +
+                                       "Student " + student.getId() + " " +
+                                       "is now taking module " + modules.getId() + "." );
+                }
+                else {
+                    System.out.println("Addition unsuccessful.\n" +
+                                       "Student " + student.getId() + " " +
+                                       "is already taking module " + modules.getId() + "." );
+                }
+            }
+            catch(NoResultException e){
+                System.out.println("There is no module with the given ID.");
+            }
+
+        }
+        catch (NoResultException e){
+            System.out.println("There is no student with the given ID.");
         }
     }
 }