+++ src/main/java/csc1035/project2/MainIO.java
@@ -290,8 +290,13 @@ public class MainIO {
             String mid = sc.nextLine();
             try {
                 Modules modules = (Modules) ic.getById( Modules.class, mid );
-                handler.removeStudentFromModule( student, modules );
-                System.out.println("Removal successful.");
+                boolean removed = handler.removeStudentFromModule( student, modules );
+                if ( removed ) {
+                    System.out.println("Removal successful.");
+                }
+                else {
+                    System.out.println("Student does not take that module.");
+                }
             }
             catch(NoResultException e){
                 System.out.println("There is no module with the given ID.");