+++ src/main/java/csc1035/project2/ModuleHandler.java
@@ -55,7 +55,7 @@ public class ModuleHandler {
      */
     public boolean addStudentToModule ( Students s, Modules m ) {
         IController ic = new Controller();
-        if( !s.getTake().contains( ic.getById( Take.class, m.getId()) ) ) {
+        if( !m.getModuleStudents().contains( s ) ) {
             Take t = new Take(s.getId(), m.getId());
 
             s.getTake().add(t);
@@ -80,7 +80,7 @@ public class ModuleHandler {
      */
     public boolean addStaffToModule ( Staff s, Modules m ) {
         IController ic = new Controller();
-        if( !s.getTeach().contains( ic.getById( Teach.class, m.getId()) ) ) {
+        if( !m.getModuleStaff().contains( s ) ) {
             Teach t = new Teach(s.getId(), m.getId());
 
             s.getTeach().add(t);
@@ -101,7 +101,7 @@ public class ModuleHandler {
      */
     public boolean removeStudentFromModule ( Students s, Modules m ) {
         IController ic = new Controller();
-        if( s.getTake().contains( ic.getById( Take.class, m.getId()) ) ) {
+        if( m.getModuleStudents().contains( s ) ) {
             s.getTake().remove(ic.getById(Take.class, m.getId()));
             refreshModuleHandler();
             return true;
@@ -118,7 +118,7 @@ public class ModuleHandler {
      */
     public boolean removeStaffFromModule ( Staff s, Modules m ) {
         IController ic = new Controller();
-        if( s.getTeach().contains( ic.getById( Teach.class, m.getId()) ) ) {
+        if( m.getModuleStaff().contains( s ) ) {
             s.getTeach().remove(ic.getById(Teach.class, m.getId()));
             refreshModuleHandler();
             return true;