+++ CHANGELOG.md
@@ -36,3 +36,4 @@
 
 ##### v0.5 [15.03.2021]: Module Handling UI
 * (ERICK) Created `moduleOptions()` method.
+* (ERICK) `addStudent/StaffToModule()` now return booleans.