+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 * (ERICK) Created `moduleOptions()` method.
 * (ERICK) `addStudent/StaffToModule()` now return booleans.
 * (ERICK) Created `studentAddModule()` method.
+* (ERICK) Created `staffAddModule` method.