+++ CHANGELOG.md
@@ -44,3 +44,4 @@
 * (ERICK) Take/teach removal checks if valid removal.
 * (ERICK) Take/teach addition checks correctly.
 * (ERICK) `studentRemoveModule()` makes use of removal check.
+* (ERICK) `staffRemoveModule()` makes use of removal check.