+++ CHANGELOG.md
@@ -45,3 +45,4 @@
 * (ERICK) Take/teach addition checks correctly.
 * (ERICK) `studentRemoveModule()` makes use of removal check.
 * (ERICK) `staffRemoveModule()` makes use of removal check.
+* (ERICK)  `menu()` no longer makes use of source 15.